Added broadcasting std::pair to utilities
